requirements.txt
streamlit>=1.32.0
requests>=2.31.0
pystray>=0.19.5
pillow>=10.0.0
yt-dlp>=2024.1.1
pytrends>=4.9.0
